refactor: change react-bootstrap import path
