--- docs/docs/auto-docs/components/test-utils/AsyncComponent/functions/default.md
@@ -1,13 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Function: default()
-
-> **default**(): `Element`
-
-Defined in: [src/components/test-utils/AsyncComponent.tsx:12](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/test-utils/AsyncComponent.tsx#L12)
-
-## Returns
-
-`Element`
