+++ docs/docs/auto-docs/test-utils/AsyncComponent/functions/default.md
@@ -1,0 +1,13 @@
+[Admin Docs](/)
+
+***
+
+# Function: default()
+
+> **default**(): `Element`
+
+Defined in: [src/test-utils/AsyncComponent.tsx:12](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/test-utils/AsyncComponent.tsx#L12)
+
+## Returns
+
+`Element`
