+++ docs/docs/auto-docs/test-utils/I18nextProviderMock/functions/I18nextProvider.md
@@ -6,7 +6,7 @@
 
 > **I18nextProvider**(`__namedParameters`): `Element`
 
-Defined in: [src/components/test-utils/I18nextProviderMock.tsx:14](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/test-utils/I18nextProviderMock.tsx#L14)
+Defined in: [src/test-utils/I18nextProviderMock.tsx:14](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/test-utils/I18nextProviderMock.tsx#L14)
 
 ## Parameters
