+++ docs/docs/auto-docs/test-utils/MockBrowserRouter/functions/default.md
@@ -6,7 +6,7 @@
 
 > **default**(`__namedParameters`): `Element`
 
-Defined in: [src/components/test-utils/MockBrowserRouter.tsx:12](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/test-utils/MockBrowserRouter.tsx#L12)
+Defined in: [src/test-utils/MockBrowserRouter.tsx:12](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/test-utils/MockBrowserRouter.tsx#L12)
 
 ## Parameters
